Started complete rewrite to include parts of Pascal Pfannes' thesis
To enable the change from three.js to babylon.js,
we had to make a lot of changes to how this package works.

We now use a proper web development setup for the web side of things,
using typescript and rollup to bundle the code. This also includes
a change to use React for the 3D scene.